# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job Search Platform API",
        "description": "API Documentation for Job Search Platform Project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/jobs": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Search and filter jobs",
                "description": "Public job search with filters for keywords, location, salary, etc.",
                "operationId": "getJobs",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search keyword in title, description, and requirements",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "description": "Filter by location",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "work_type",
                        "in": "query",
                        "description": "Filter by work type (Full-time, Part-time, Contract, Internship)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "workplace_type",
                        "in": "query",
                        "description": "Filter by workplace type (Remote, On-site, Hybrid)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "salary_min",
                        "in": "query",
                        "description": "Minimum salary",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "salary_max",
                        "in": "query",
                        "description": "Maximum salary",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "company_id",
                        "in": "query",
                        "description": "Filter by company ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "skill_ids",
                        "in": "query",
                        "description": "Comma-separated skill IDs",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "industry",
                        "in": "query",
                        "description": "Filter by industry (company field of work)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort order: latest (default), salary_desc, salary_asc",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "latest",
                                "salary_desc",
                                "salary_asc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (default 15, max 50)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of jobs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/jobs/{id}": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Get job details",
                "description": "Returns detailed information about a specific job.",
                "operationId": "getJobDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job details"
                    },
                    "404": {
                        "description": "Job not found"
                    }
                }
            }
        },
        "/favorites/{jobId}": {
            "post": {
                "tags": [
                    "Jobs",
                    "Favorites"
                ],
                "summary": "Add job to favorites",
                "description": "Add job to favorites (for job seekers).",
                "operationId": "addFavoriteJob",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job saved to favorites"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Jobs",
                    "Favorites"
                ],
                "summary": "Remove job from favorites",
                "description": "Remove job from favorites.",
                "operationId": "removeFavoriteJob",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job removed from favorites"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/favorites": {
            "get": {
                "tags": [
                    "Jobs",
                    "Favorites"
                ],
                "summary": "Get favorite jobs",
                "description": "Get user's favorite jobs.",
                "operationId": "getFavoriteJobs",
                "responses": {
                    "200": {
                        "description": "List of favorite jobs"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/jobs/matching": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Get matching jobs",
                "description": "Returns jobs that match the active or specified CV of the job seeker.",
                "operationId": "getMatchingJobs",
                "parameters": [
                    {
                        "name": "cv_id",
                        "in": "query",
                        "description": "CV ID for matching. If missing, uses the latest CV.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "min_score",
                        "in": "query",
                        "description": "Minimum match score (default 50)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of matching jobs"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/employer/jobs/{id}/publish": {
            "post": {
                "tags": [
                    "Employer",
                    "Jobs"
                ],
                "summary": "Publish a job",
                "description": "Publish a job (change status to Active).",
                "operationId": "publishJob",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job published"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/employer/jobs/{id}/close": {
            "post": {
                "tags": [
                    "Employer",
                    "Jobs"
                ],
                "summary": "Close a job",
                "description": "Close a job (change status to Closed).",
                "operationId": "closeJob",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job closed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}